+++ packages/example/CHANGELOG.md
@@ -1,5 +1,88 @@
 # example
 
+## 0.0.46
+
+### Patch Changes
+
+- Updated dependencies [924ae82]
+- Updated dependencies [5e233ea]
+  - @rainbow-me/rainbowkit@0.11.1
+
+## 0.0.45
+
+### Patch Changes
+
+- Updated dependencies [1876ba0]
+  - @rainbow-me/rainbowkit@0.11.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.8
+
+## 0.0.44
+
+### Patch Changes
+
+- Updated dependencies [355402b]
+  - @rainbow-me/rainbowkit@0.10.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.7
+
+## 0.0.43
+
+### Patch Changes
+
+- Updated dependencies [49f0ec9]
+  - @rainbow-me/rainbowkit@0.9.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.6
+
+## 0.0.42
+
+### Patch Changes
+
+- Updated dependencies [a1d6776]
+  - @rainbow-me/rainbowkit@0.8.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.5
+
+## 0.0.41
+
+### Patch Changes
+
+- Updated dependencies [6b37050]
+- Updated dependencies [6b37050]
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.5
+  - @rainbow-me/rainbowkit@0.8.0
+
+## 0.0.40
+
+### Patch Changes
+
+- Updated dependencies [e36da59]
+- Updated dependencies [0ff4210]
+  - @rainbow-me/rainbowkit@0.7.4
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
+
+## 0.0.39
+
+### Patch Changes
+
+- Updated dependencies [5a65178]
+  - @rainbow-me/rainbowkit@0.7.3
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
+
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [1de8203]
+  - @rainbow-me/rainbowkit@0.7.2
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
+
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [6b6a73b]
+- Updated dependencies [5ddc813]
+  - @rainbow-me/rainbowkit@0.7.1
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.4
+
 ## 0.0.36
 
 ### Patch Changes
@@ -76,7 +159,7 @@
 - Updated dependencies [737a1d6]
 - Updated dependencies [737a1d6]
 - Updated dependencies [488c5a1]
-  - @rainbow-me/rainbowkit-siwe-next-auth@1.0.0
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.6
   - @rainbow-me/rainbowkit@0.5.0
 
 ## 0.0.28